Changed arguments of ‘history_log()’ and ‘program()’, as well as rearranging the declarations
Changed both ‘history_log()’ and ‘program()’ to accept a ‘History_Log’ object. The object is passed by reference to ensure that this object remains the same through multiple runs of this program. Also moved the declarations for ‘History_Log.cpp’ to be above ‘Calculator.cpp’ in order for this change in argument to work.