Added the ‘previous answer memory’ feature and updated the ‘History  Log’ section under ‘Help’
Modified ‘History Log’ to have a ‘previous answer memory’ feature where the user will be able to view their most recent answer. The ‘History Log’ section in ‘Help’ has also been updated to instruct the user on how to use this feature when inputting an expression. Currently, the calculator can’t understand this feature, but it will be updated in the future.