Added the ability to use the previous answer (Ans) in an expression
Reworked the logic of calculate() to be able to accept ‘Ans/ans’ (previous answer memory) as a valid digit input, using the new helper function ‘convert_to_digits()’. Additionally, with this rework, the program can catch more errors of incorrect syntax in the inputted expression. The program now also recognizes a single number inputted as a valid expression.