‘clear()’ function modified to be able to delete nodes from a custom interval
Given a start and end index, the function will delete nodes at ‘start’ and end at ‘end’ inclusively. The new pointer, ‘_marker’ helps to reattach the linked list if necessary, and ‘_tail’ is reassigned to its proper position based on the state of the list after deletion. The bool argument is only used for preventing the destructor from outputting the error message about no memory.