Clearing memory feature has been added and ‘History Log’ under ‘Help’ is updated
The user is able to clear memory now, and is given the option to clear all at once or a specific interval. If selecting a specific inclusive interval, the user will be prompted to select the start and end points. If valid, the program will delete the specified interval and reassign the tail if necessary, so ‘previous answer memory’ will be updated properly. The ‘History Log’ category under ‘Help’ has also been updated accordingly to help the user understand how to use this new feature.